+++ crates/fuel-core/src/executor.rs
@@ -41,7 +41,7 @@ mod tests {
         },
         entities::{
             coins::coin::CompressedCoin,
-            message::{
+            relayer::message::{
                 Message,
                 MessageV1,
             },
@@ -2954,7 +2954,10 @@ mod tests {
             },
             StorageAsMut,
         };
-        use fuel_core_types::fuel_merkle::binary::root_calculator::MerkleRootCalculator;
+        use fuel_core_types::{
+            entities::RelayedTransaction,
+            fuel_merkle::binary::root_calculator::MerkleRootCalculator,
+        };
 
         fn database_with_genesis_block(da_block_height: u64) -> Database<OnChain> {
             let mut db = Database::default();
@@ -2972,6 +2975,16 @@ mod tests {
             let da_height = message.da_height();
             db.storage::<EventsHistory>()
                 .insert(&da_height, &[Event::Message(message)])
+                .expect("Should insert event");
+        }
+
+        fn add_events_to_relayer(
+            db: &mut Database<Relayer>,
+            da_height: DaBlockHeight,
+            events: &[Event],
+        ) {
+            db.storage::<EventsHistory>()
+                .insert(&da_height, events)
                 .expect("Should insert event");
         }
 
@@ -3112,11 +3125,23 @@ mod tests {
             let relayer_da_height = 10u64;
             let mut root_calculator = MerkleRootCalculator::new();
             for da_height in (genesis_da_height + 1)..=relayer_da_height {
+                // message
                 let mut message = Message::default();
                 message.set_da_height(da_height.into());
                 message.set_nonce(da_height.into());
                 root_calculator.push(message.id().as_ref());
-                add_message_to_relayer(&mut relayer_db, message);
+                // transaction
+                let mut transaction = RelayedTransaction::default();
+                transaction.set_da_height(da_height.into());
+                transaction.set_max_gas(da_height);
+                transaction.set_serialized_transaction(da_height.to_be_bytes().to_vec());
+                root_calculator.push(Bytes32::from(transaction.id()).as_ref());
+                // add events to relayer
+                add_events_to_relayer(
+                    &mut relayer_db,
+                    da_height.into(),
+                    &[message.into(), transaction.into()],
+                );
             }
             let producer = create_relayer_executor(on_chain_db, relayer_db);
             let block = test_block(block_height.into(), relayer_da_height.into(), 0);